# admin/src/main/java/com/base/modules/sys/controller/SysMenuController.java
package com.base.modules.sys.controller;

import com.base.common.annotation.SysLog;
import com.base.common.exception.RRException;
import com.base.common.utils.Constant;
import com.base.common.utils.R;
import com.base.modules.sys.entity.SysMenuEntity;
import com.base.modules.sys.service.SysMenuService;
import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import org.apache.commons.lang.StringUtils;
import org.apache.shiro.authz.annotation.RequiresPermissions;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;

import java.util.List;

/**
 * 系统菜单
 * 
 * @date 2016年10月27日 下午9:58:15
 */
@RestController
@RequestMapping("/sys/menu")
@Api(tags="后台菜单管理")
public class SysMenuController extends AbstractController {
	@Autowired
	private SysMenuService sysMenuService;

	/**
	 * 导航菜单
	 */
	@GetMapping("/nav")
	public R nav(){
		List<SysMenuEntity> menuList = sysMenuService.getUserMenuList(getUserId());
		return R.ok().put("menuList", menuList);
	}
	/**
	 * 所有菜单列表
	 */
	@PostMapping("/list")
	@RequiresPermissions("sys:menu:list")
	@ApiOperation("列表菜单查询")
	public List<SysMenuEntity> list(@RequestBody SysMenuEntity sysMenuEntity){
		List<SysMenuEntity> menuList = sysMenuService.querySysMenuEntityListByUserId(sysMenuEntity,getUserId());
		return menuList;
	}
	/**
	 * 不包含前端网站菜单的菜单列表
	 */
	@RequestMapping("/notFrontlist")
	@RequiresPermissions("sys:menu:list")
	public List<SysMenuEntity> notFrontlist(){
		List<SysMenuEntity> menuList = sysMenuService.queryNotFrontMenuList();
		for(SysMenuEntity sysMenuEntity : menuList){
			SysMenuEntity parentMenuEntity = sysMenuService.selectById(sysMenuEntity.getParentId());
			if(parentMenuEntity != null){
				sysMenuEntity.setParentName(parentMenuEntity.getName());
			}
		}

		return menuList;
	}

	/**
	 * 根据用户获取，当前用户的菜单（不包含前端网站菜单的菜单列表）
	 */
	@GetMapping("/getUserMenuList")
	public List<SysMenuEntity> queryNotFrontMenuList(){
		List<SysMenuEntity> menuList = sysMenuService.queryNotFrontMenuList(getUserId());
		return menuList;
	}

	
	/**
	 * 选择菜单(添加、修改菜单)
	 */
	@GetMapping("/select")
	@RequiresPermissions("sys:menu:select")
	@ApiOperation(value="选择菜单(添加、修改菜单)")
	public R select(SysMenuEntity sysMenuEntity){
		//查询列表数据
		List<SysMenuEntity> menuList = sysMenuService.queryNotButtonList(sysMenuEntity);
		
		//添加顶级菜单
		SysMenuEntity root = new SysMenuEntity();
		root.setMenuId(0L);
		root.setName("一级菜单");
		root.setParentId(-1L);
		root.setOpen(true);
		menuList.add(root);
		
		return R.ok().put("menuList", menuList);
	}
	
	/**
	 * 菜单信息
	 */
	@RequestMapping("/info/{menuId}")
	@RequiresPermissions("sys:menu:info")
	public R info(@PathVariable("menuId") Long menuId){
		SysMenuEntity menu = sysMenuService.selectById(menuId);
		return R.ok().put("menu", menu);
	}
	
	/**
	 * 保存
	 */
	@SysLog("保存菜单")
	@RequestMapping("/save")
//    @CustomizeCacheAnnotation(value="删除字典触发缓存",moduleId = "9", type = "delete") //目的是为触发完网站导航信息，与后端无关系
	@RequiresPermissions("sys:menu:save")
	public R save(@RequestBody SysMenuEntity menu){
		//前台不支持的徐选择的情况下，默认加的菜单为后台菜单，默认显示
//		menu.setKind(0);
		menu.setIsenable("Y");
		//数据校验
		verifyForm(menu);
		
		sysMenuService.insert(menu);
		
		return R.ok();
	}
	
	/**
	 * 修改
	 */
	@SysLog("修改菜单")
	@RequestMapping("/update")
//	@CustomizeCacheAnnotation(value="删除字典触发缓存",moduleId = "9", type = "delete") //目的是为触发完网站导航信息，与后端无关系
	@RequiresPermissions("sys:menu:update")
	public R update(@RequestBody SysMenuEntity menu){
		//数据校验
		verifyForm(menu);
				
		sysMenuService.updateById(menu);
		
		return R.ok();
	}
	
	/**
	 * 删除
	 */
	@SysLog("删除菜单")
	@RequestMapping("/delete")
//	@CustomizeCacheAnnotation(value="删除字典触发缓存",moduleId = "9", type = "delete") //目的是为触发完网站导航信息，与后端无关系
	@RequiresPermissions("sys:menu:delete")
	public R delete(long menuId){
		if(menuId <= 31){
			return R.error("系统菜单，不能删除");
		}

		//判断是否有子菜单或按钮
		List<SysMenuEntity> menuList = sysMenuService.queryListParentId(menuId);
		if(menuList.size() > 0){
			return R.error("请先删除子菜单或按钮");
		}

		sysMenuService.delete(menuId);

		return R.ok();
	}
	
	/**
	 * 验证参数是否正确
	 */
	private void verifyForm(SysMenuEntity menu){
		if(StringUtils.isBlank(menu.getName())){
			throw new RRException("菜单名称不能为空");
		}
		
		if(menu.getParentId() == null){
			throw new RRException("上级菜单不能为空");
		}
		
		//菜单
		if(menu.getType() == Constant.MenuType.MENU.getValue()){
			if(StringUtils.isBlank(menu.getUrl())){
				throw new RRException("菜单URL不能为空");
			}
		}
		
		//上级菜单类型
		int parentType = Constant.MenuType.CATALOG.getValue();
		if(menu.getParentId() != 0){
			SysMenuEntity parentMenu = sysMenuService.selectById(menu.getParentId());
			parentType = parentMenu.getType();
		}
		
		//目录、菜单
		if(menu.getType() == Constant.MenuType.CATALOG.getValue() ||
				menu.getType() == Constant.MenuType.MENU.getValue()){
			if(parentType != Constant.MenuType.CATALOG.getValue()){
				throw new RRException("上级菜单只能为目录类型");
			}
			return ;
		}
		
		//按钮
		if(menu.getType() == Constant.MenuType.BUTTON.getValue()){
			if(parentType != Constant.MenuType.MENU.getValue()){
				throw new RRException("上级菜单只能为菜单类型");
			}
			return ;
		}
	}
}